Fix(Reports):fix reports shows as my-reports , reports and fix leaderboard and fix 404
